Explicitly set spellcheck to false
Unfortunately when custom solr parameters are passed to the blacklight
get_search_results function, it overrides some of the sane defaults that
are set. One, notably, is setting spellcheck to false. Because this
isn't specified, and blacklight also automatically sets spellcheck.q to
be the value of q, we get errors now that we have spellcheck disabled.